refactor(models): rename model-colors.ts to consts.ts
