+++ apps/sim/app/(landing)/models/components/model-timeline-chart.tsx
@@ -2,7 +2,7 @@
 
 import { useMemo } from 'react'
 import Link from 'next/link'
-import { getProviderColor } from '@/app/(landing)/models/components/model-colors'
+import { getProviderColor } from '@/app/(landing)/models/components/consts'
 import type { CatalogModel } from '@/app/(landing)/models/utils'
 
 function formatShortDate(date: string): string {
